requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
macholib==1.16.3
packaging==23.2
pillow==10.2.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
